V 1.4
- updated UI
- moved code from MainActivity
