+++ app/src/main/res/drawable/background_fade.xml
@@ -3,7 +3,7 @@
     android:shape="rectangle">
     <gradient
         android:angle="90"
-        android:startColor="#000000"
-        android:endColor="#0955A6"
+        android:startColor="#2196F3"
+        android:endColor="#05243D"
         android:type="linear" />
 </shape>
